split tcp_socket file
