+++ examples/simple_example.cc
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 #include <ntio/sockets/tcp_socket.h>
+#include <ntio/sockets/tcp_listener.h>
 
 #include <coroutine>
 #include <iostream>
