+++ src/ntio/sockets/tcp_socket.cc
@@ -46,30 +46,4 @@ AsyncWrite TcpSocket::Write(const std::vector<uint8_t>& data, size_t len) const 
   return AsyncWrite(fd_, settings_.write_timeout, data, len);
 }
 
-/*
- * TcpListener
- */
-
-const TcpListener::Settings TcpListener::kDefaultSettings = {
-    .back_log = 10,
-    .accept_timeout = std::chrono::milliseconds(1000),
-};
-
-TcpListener::TcpListener() : TcpListener(kDefaultSettings) {}
-
-TcpListener::TcpListener(TcpListener::Settings settings) : fd_(OpenTcpSocket()), settings_(settings) {
-  LOG_DEBUG("Opened socket listen_fd=" + std::to_string(fd_));
-}
-
-TcpListener::~TcpListener() {
-  LOG_DEBUG("Close socket listen_fd=" + std::to_string(fd_));
-  CloseTcpSocket(fd_);
-}
-
-AsyncBind TcpListener::Bind(EndPoint end_point) const noexcept {
-  return AsyncBind(fd_, std::move(end_point), settings_.back_log);
-}
-
-AsyncAccept TcpListener::Accept() const noexcept { return AsyncAccept(fd_, settings_.accept_timeout); }
-
 }  // namespace ntio::sockets
